+++ src/brida/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Stable Brida command-line entrypoints."""